applet: Move the trick logic to applet package
