+++ src/rprocessing/applet/BuiltinApplet.java
@@ -1,0 +1,23 @@
+package rprocessing.applet;
+
+import processing.core.PApplet;
+public class BuiltinApplet extends PApplet {
+    /*
+     * Wrapper functions
+     * 
+     * float is not implemented in R, so processing.r need to cast double to
+     * float.
+     */
+
+    public void size(double width, double height) {
+        super.size((int) width, (int) height);
+    }
+
+    public void point(double x, double y) {
+        super.point((float) x, (float) y);
+    }
+
+    public void line(double posAX, double posAY, double posBX, double posBY) {
+        super.line((float) posAX, (float) posAY, (float) posBX, (float) posBY);
+    }
+}
